Make `WeakRefVirtualProxy` to separate file
